ProgramacionIVContenido/tareas/tarea_4_Migracion_a_Base_de_Datos_No_Relacional_con_MongoDB/requirements.txt
pymongo>=4.7.0
python-dotenv>=1.0.1
